The black hole actually consumes the planets
Earlier the planets and orbits were visible inside the black hole. To
fix this, compare the distance of the planet from the sun's original
position with the current radius of the sun. If the black hole has
covered the planet, this distance will be lesser than the current sun
radius. In this case, don't display the plannets.